Fixed multilanguage support in forms for questions
