+++ app/models/question.rb
@@ -12,7 +12,7 @@
 
 class Question < ApplicationRecord
   has_many :answers, as: :answerable
-  has_many :translations
-  accepts_nested_attributes_for :translations
+
   translates :text
+  accepts_nested_attributes_for :translations
 end
