+++ db/schema.rb
@@ -92,7 +92,6 @@ ActiveRecord::Schema.define(version: 20180208152748) do
     t.datetime "updated_at", null: false
     t.string "uuid", null: false
     t.index ["user_id"], name: "index_surveys_on_user_id"
-    t.index ["uuid"], name: "index_surveys_on_uuid", unique: true
   end
 
   create_table "users", force: :cascade do |t|
