+++ test/models/survey_test.rb
@@ -13,7 +13,6 @@
 # Indexes
 #
 #  index_surveys_on_user_id  (user_id)
-#  index_surveys_on_uuid     (uuid) UNIQUE
 #
 # Foreign Keys
 #
